Move queryForFacilities tests over
